Move decompositions to op_math (#4906)
**Description of the Change:**
Move the operators/unitary decompositions to the op_math module and
update the documentation. Also moved the tests.

---------
