+++ pennylane/ops/op_math/controlled_decompositions.py
@@ -188,9 +188,7 @@ def ctrl_decomp_zyz(target_operation: Operator, control_wires: Wires):
         phi, theta, omega = target_operation.single_qubit_rot_angles()
     except NotImplementedError:
         with qml.QueuingManager.stop_recording():
-            zyz_decomp = qml.transforms.one_qubit_decomposition(
-                qml.matrix(target_operation), target_wire
-            )
+            zyz_decomp = qml.ops.one_qubit_decomposition(qml.matrix(target_operation), target_wire)
         phi, theta, omega = [gate.parameters[0] for gate in zyz_decomp]
 
     decomp = []
