+++ pennylane/transforms/decompositions/__init__.py
@@ -11,11 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-r"""This module contains decompositions for (numerically-specified) arbitrary
-unitary operations into sequences of elementary operations.
+r"""This module contains decompositions for quantum circuits.
 """
 
+from .clifford_t_transform import clifford_t_decomposition
 from .single_qubit_unitary import one_qubit_decomposition
 from .two_qubit_unitary import two_qubit_decomposition
-
-from .clifford_t import clifford_t_decomposition, sk_decomposition
