+++ tests/transforms/test_cliffordt_transform.py
@@ -19,7 +19,7 @@ from functools import reduce
 import pytest
 import pennylane as qml
 
-from pennylane.transforms.decompositions.clifford_t.clifford_t_transform import (
+from pennylane.transforms.decompositions.clifford_t_transform import (
     check_clifford_t,
     clifford_t_decomposition,
     _rot_decompose,
